# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Meo Mai Moi API",
        "description": "API for Meo Mai Moi, a cat care management platform.",
        "contact": {
            "email": "support@meomaimoi.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Development server"
        }
    ],
    "paths": {
...
        "/api/email/verify/{id}/{hash}": {
            "get": {
                "tags": [
                    "Email Verification"
                ],
                "summary": "Verify email address",
                "description": "Verify user's email address using signed URL from verification email.",
                "operationId": "007cd3d5c6498af4e175e24329cae026",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "hash",
                        "in": "path",
                        "description": "Email hash",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "expires",
                        "in": "query",
                        "description": "Expiration timestamp",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "signature",
                        "in": "query",
                        "description": "URL signature",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Email verified successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Email verified successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Email already verified or invalid request"
                    },
                    "403": {
                        "description": "Invalid or expired verification link"
                    }
                }
            }
        },
...
        "/helper-profiles/{id}": {
            "get": {
                "tags": [
                    "Helper Profiles"
                ],
                "summary": "Get a specific helper profile",
                "operationId": "3fff3617c2915451e3e7c237bd331b06",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the helper profile",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The helper profile",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HelperProfile"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Helper profile not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Helper Profiles"
                ],
                "summary": "Update a helper profile",
                "operationId": "809e264b7c665d8c88d5c485af9f3fe5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the helper profile",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/HelperProfile"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Helper profile updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HelperProfile"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Helper Profiles"
                ],
                "summary": "Update a helper profile (DEPRECATED - use PUT instead)",
                "description": "This endpoint is deprecated. Use PUT /helper-profiles/{id} instead. Kept for HTML form compatibility.",
                "operationId": "f932788c0c3517bbe5ec4ba59c9cc574",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the helper profile",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/HelperProfile"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Helper profile updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HelperProfile"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "deprecated": true
            },
            "delete": {
                "tags": [
                    "Helper Profiles"
                ],
                "summary": "Delete a helper profile",
                "operationId": "945ba6975829c4c07247fd546121f396",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the helper profile",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Helper profile deleted successfully"
                    },
                    "400": {
                        "description": "Cannot delete profile with associated placement requests"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HelperProfile": {
                "title": "HelperProfile",
                "description": "Helper Profile model",
                "properties": {
                    "id": {
                        "description": "Helper Profile ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "user_id": {
                        "description": "ID of the associated user",
                        "type": "integer",
                        "format": "int64"
                    },
                    "status": {
                        "description": "Status of the helper profile",
                        "type": "string",
                        "enum": [
                            "active",
                            "archived",
                            "deleted"
                        ]
                    },
                    "request_types": {
                        "description": "Types of placement requests this helper can respond to",
                        "type": "array",
                        "items": {
                            "type": "string",
                            "enum": [
                                "foster_paid",
                                "foster_free",
                                "permanent"
                            ]
                        }
                    },
                    "country": {
                        "description": "ISO 3166-1 alpha-2 country code",
                        "type": "string"
                    },
                    "state": {
                        "description": "State/Province",
                        "type": "string",
                        "nullable": true
                    },
                    "city": {
                        "description": "City",
                        "type": "string",
                        "nullable": true
                    },
                    "address": {
                        "description": "Street address",
                        "type": "string",
                        "nullable": true
                    },
                    "zip_code": {
                        "description": "ZIP/Postal code",
                        "type": "string",
                        "nullable": true
                    },
                    "contact_info": {
                        "description": "Additional contact information visible to pet owners when responding to placement requests",
                        "type": "string",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Timestamp of helper profile creation",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Timestamp of last helper profile update",
                        "type": "string",
                        "format": "date-time"
                    },
                    "archived_at": {
                        "description": "Timestamp when the profile was archived",
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "restored_at": {
                        "description": "Timestamp when the profile was restored",
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "deleted_at": {
                        "description": "Timestamp when the profile was deleted",
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}